Migrate to new Adafruit Library (#18)
* First draft of changing all references (untested)

* Adaption to new API

* Fix unsmooth pixles

The effects were very unsmooth and bad, this happened, because the pixel values were directly written. Now, we turn off auto_write and wait until we call show().

* Some beautifictions

* Remove old files

I don't think that we need those files anymore, and even if we do, we can look them up in the history.