Major version upgrade of flutter and dependencies
This upgrades the whole codebase to the most current flutter and dart version, along with all dependencies.
This results in a slightly more modern look and feel, many code refactorings to handle deprecations and a new look for the dark theme.
It should now also be possible to build the app with newer versions of dev tools. For this, many underlying files have also been migrated